feat: make package pip-installable
